+++ CHANGELOG.md
@@ -4,8 +4,15 @@
 ### Summary
 * Started CHANGELOG.md (but for now it is far from definitive, but is intended to include all breaking changes)
 * Added support for containers, documented in Custom Form Schemas section of documentation
+
 ### BREAKING CHANGES
-* Changes form-input directive so that it expects <form-input schema="formSchema"> instead of
-<form-input ng-repeat="field in formSchema">
+* Changes form-input directive so that it expects
+```
+    <form-input schema="formSchema">
+```    
+instead of
+```
+    <form-input ng-repeat="field in formSchema" info={{field}}>
+```    
 * Bespoke directives need to be modified (changes will normally be from something like
  var info = JSON.parse(attrs.info); to var info = scope[attrs.schema][0];